requirements.txt
streamlit
speech_recognition
google.generativeai
dotenv
os
json
anthropic
openai